requirements.txt
pandas
numpy
scikit-learn
joblib
requests
Flask
flask-socketio
fastapi
uvicornannotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
autobahn==19.11.2
Automat==24.8.1
beautifulsoup4==4.13.3
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dateutils==0.6.12
fastapi==0.115.8
Flask==3.1.0
Flask-SocketIO==5.5.1
fonttools==4.56.0
frozendict==2.4.6
h11==0.14.0
html5lib==1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiteconnect==5.0.1
kiwisolver==1.4.8
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.10.0
multitasking==0.0.11
nsetools==1.0.11
numpy==2.2.2
packaging==24.2
pandas==2.2.3
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.1
pywin32==308
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
service-identity==24.2.0
setuptools==75.8.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.3
ta==0.11.0
threadpoolctl==3.5.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
webencodings==0.5.1
websockets==14.2
Werkzeug==3.1.3
wsproto==1.2.0
yfinance==0.2.52
zope.interface==7.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
autobahn==19.11.2
Automat==24.8.1
beautifulsoup4==4.13.3
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dateutils==0.6.12
fastapi==0.115.8
Flask==3.1.0
Flask-SocketIO==5.5.1
fonttools==4.56.0
frozendict==2.4.6
future==1.0.0
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiteconnect==5.0.1
kiwisolver==1.4.8
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.10.0
multitasking==0.0.11
nsetools==1.0.11
numpy==2.2.3
packaging==24.2
pandas==2.2.3
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.6
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.1
pywin32==308
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
service-identity==24.2.0
setuptools==75.8.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.45.3
ta==0.11.0
threadpoolctl==3.5.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
upstox==2.0.3
upstox-python-sdk==2.12.0
urllib3==2.3.0
uuid==1.30
uvicorn==0.34.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wsproto==1.2.0
yfinance==0.2.52
zope.interface==7.2
